+++ src/PostgREST/Middleware.hs
@@ -1,5 +1,6 @@
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE FlexibleContexts #-}
 
 module PostgREST.Middleware where
 
@@ -15,10 +16,11 @@ import           Network.Wai.Middleware.Gzip   (def, gzip)
 import           Network.Wai.Middleware.Static (only, staticPolicy)
 
 import           PostgREST.ApiRequest          (ApiRequest(..))
-import           PostgREST.Auth                (claimsToSQL, JWTAttempt(..))
+import           PostgREST.Auth                (JWTAttempt(..))
 import           PostgREST.Config              (AppConfig (..), corsPolicy)
 import           PostgREST.Error               (simpleError)
 import           PostgREST.Types               (ContentType (..), toHeader)
+import           PostgREST.QueryBuilder        (pgFmtIdent, pgFmtLit, unquoted)
 
 import           Protolude                     hiding (concat, null)
 
@@ -31,14 +33,19 @@ runWithClaims conf eClaims app req =
     JWTInvalid -> return $ unauthed "JWT invalid"
     JWTMissingSecret -> return $ simpleError status500 "Server lacks JWT secret"
     JWTClaims claims -> do
-      -- role claim defaults to anon if not specified in jwt
-      let setClaims = claimsToSQL (M.union claims (M.singleton "role" anon))
-      H.sql $ mconcat setClaims
+      H.sql $ mconcat $ setRoleSql ++ claimsSql ++ headersSql
       mapM_ H.sql customReqCheck
       app req
+      where
+        headersSql = envVarToSql "request.header." $ iHeaders req
+        claimsSql = envVarToSql "request.jwt.claim." [(c,unquoted v) | (c,v) <- M.toList claimsWithRole]
+        setRoleSql = maybeToList $
+          (\r -> "set local role " <> r <> ";") . toS . pgFmtLit . unquoted <$> M.lookup "role" claimsWithRole
+        -- role claim defaults to anon if not specified in jwt
+        claimsWithRole = M.union claims (M.singleton "role" anon)
+        anon = String . toS $ configAnonRole conf
+        customReqCheck = (\f -> "select " <> toS f <> "();") <$> configReqCheck conf
   where
-    anon = String . toS $ configAnonRole conf
-    customReqCheck = (\f -> "select " <> toS f <> "();") <$> configReqCheck conf
     unauthed message = responseLBS unauthorized401
       [ toHeader CTApplicationJSON
       , ( "WWW-Authenticate"
@@ -47,6 +54,13 @@ runWithClaims conf eClaims app req =
         )
       ]
       (toS $ "{\"message\":\""<>message<>"\"}")
+
+envVarToSql :: Text -> [(Text, Text)] -> [ByteString]
+envVarToSql prefix vars = varStmts
+ where
+  varStmts = map setVar vars
+  setVar (k, val) = toS $ "set local " <> pgFmtIdent (prefix <> k)
+                    <> " = " <> pgFmtLit val <> ";"
 
 defaultMiddle :: Application -> Application
 defaultMiddle =